pythonProject5/requirements.txt
pymongo~=4.3.3
fastapi~=0.95.1
pydantic~=1.10.7
uvicorn~=0.22.0
python-dotenv~=1.0.0
paho-mqtt~=1.6.1